Remove global workers -- workers are no per proxy.
Btw, with the most recent change, each proxy has its own REQ/REP
and SURVEY sockets, not shared with anything else.  This will make
it easier to match the incoming requests to the proxy, and therefore
apply roles and limit permissions accordingly.

We should probably harden this abit as well, by adding a pipe callback
on the REP pipes, to reject connections that are not coming from the
expected proxy node.