Controller logic now handled like proxies.
The controllers are tied to the configuration now.  We also
have registration for worker ops (really controller ops would have
been a better name) and versioning, which should make it easier to
for future extensions and facilitates validation of both the ops
vector and the "type" listed in the configuration file.