Authentication token changes.
This adds a method to retrieve the full value of an authentication
token.  It also changes the "expire" member of the token to "expires".
We have added the creation time stamp as "created" as well, which
may help with auditing. Finally, the retrieval and create methods
now return the list of roles which were actually granted for a token.