Controller names could be shorter, and ops could self-name.
We have eliminated the "controller_" prefix from the controller
types in the configuration file for workers, and now the controller
ops vector provides its own name.

If we could rely on automatic initilization code via ELF, we
could actually make the controllers register themselves.  Unfortunately
the method of doing this is not portable, so for now the worker.c
file still has to list them all, but it only needs to know one
pointer, not all of them.