Add builtin roles, and fix handling of token-derived roles.
%admin - granted only explicitly, holders have all superpowers.
%all - everyone has this role implicitly.
%token - granted to users who have come via token

The %token role is intended to be used in "deny" blocks to prevent
API users from using tokens to create subtokens.

The %token and %all roles cannot be granted explicitly, or named
explicitly in a token.  (%admin *can* however.)

User defined token names now have much looser rules on their
names, as only the first character needs to be an alphanumeric or
underscore.  The remaining characters can be anything printable.

Effective role changes were not being applied to API tokens, which
ignored proxy limitations.  That is now fixed.

The signature of get_auth_perms() is now somewhat simplified.