Added generic proxy RPC method
POST /api/1.0/proxy/<name>/rpc/<method>

Proxy will forward the RCP request to the controller specified
by <name> using the <method> and the body as params.

This will reduce the need to translate every call from REST
to RPC at the proxy, as the client can use rpc for calls that
do not need a REST representation.