Role groups, and fix bugs in the role bitmasks.
Its now possible to specify rolegroups, which can be thought of
as macros to be used in place of single role names, but which can
combine multiple roles together.

It's even possible for a rolegroup to include a previously defined
rolegroup.  So you can nest provided you get the order correct.

I've cleaned up the parsing logic a bit as well, and fixed a bug
that prevented roles from actually being assigned a mask.