Initial swag at server restart.
This is fragile right now, and the timing changes in the server
have made it more so.  However, sometimes it works.  I suspect
that there are some defects in the ZeroTier transport relating to
request routing, or possibly node reuse, as what occurs is that
the proxy no longer gets survey responses when this happens.

I believe this is probably an NNG zerotier bug.