Worker should treat requests as idempotent.
We use the request ID, and resend replies, from requests that
were replayed within 30 seconds.  (This probably could be made tunable).

Responses are held for up to 30 seconds; a reaper thread is used to clean
up stale responses.  The reaper wakes every 10 seconds as long as there are
any requests pending or still in the log.  (Once the log is fully empty,
the reaper will go to sleep indefinitely, until another request comes in,
thereby letting the system go into deep sleep.)